external/asf4: Upgrade to latest version
A new project was created in Atmel Start (start.atmel.com) and all
drivers we are using was added and our clock configuration was setup.

After that the content of the asf4-drivers directory was removed and the
new project was unpacked in its place.

Finally the qtouch and sdcard drivers were reverted as they have been
modified by us.

NOTE: Some code was present in the project that wasn't used, like an
USART and an AES driver.

NOTE: The ASF4 drivers doesn't take `const` pointers when writing
buffers, so we need to cast from `const` to non-`const`. This is cought
by GCC, so the check was removed from clang-tidy.

NOTE: Moved some of our additions to the `src` directory. Like
sd_mmc_ext.c

+++ external/asf4-drivers/hri/hri_rstc_d51.h
@@ -3,41 +3,32 @@
  *
  * \brief SAM RSTC
  *
- * Copyright (C) 2016 Atmel Corporation. All rights reserved.
+ * Copyright (c) 2016-2018 Microchip Technology Inc. and its subsidiaries.
  *
  * \asf_license_start
  *
  * \page License
  *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
+ * Subject to your compliance with these terms, you may use Microchip
+ * software and any derivatives exclusively with Microchip products.
+ * It is your responsibility to comply with third party license terms applicable
+ * to your use of third party software (including open source software) that
+ * may accompany Microchip software.
  *
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. The name of Atmel may not be used to endorse or promote products derived
- *    from this software without specific prior written permission.
- *
- * 4. This software may only be redistributed and used in connection with an
- *    Atmel microcontroller product.
- *
- * THIS SOFTWARE IS PROVIDED BY ATMEL "AS IS" AND ANY EXPRESS OR IMPLIED
- * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT ARE
- * EXPRESSLY AND SPECIFICALLY DISCLAIMED. IN NO EVENT SHALL ATMEL BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS
- * OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
- * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
- * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
+ * THIS SOFTWARE IS SUPPLIED BY MICROCHIP "AS IS". NO WARRANTIES,
+ * WHETHER EXPRESS, IMPLIED OR STATUTORY, APPLY TO THIS SOFTWARE,
+ * INCLUDING ANY IMPLIED WARRANTIES OF NON-INFRINGEMENT, MERCHANTABILITY,
+ * AND FITNESS FOR A PARTICULAR PURPOSE. IN NO EVENT WILL MICROCHIP BE
+ * LIABLE FOR ANY INDIRECT, SPECIAL, PUNITIVE, INCIDENTAL OR CONSEQUENTIAL
+ * LOSS, DAMAGE, COST OR EXPENSE OF ANY KIND WHATSOEVER RELATED TO THE
+ * SOFTWARE, HOWEVER CAUSED, EVEN IF MICROCHIP HAS BEEN ADVISED OF THE
+ * POSSIBILITY OR THE DAMAGES ARE FORESEEABLE.  TO THE FULLEST EXTENT
+ * ALLOWED BY LAW, MICROCHIP'S TOTAL LIABILITY ON ALL CLAIMS IN ANY WAY
+ * RELATED TO THIS SOFTWARE WILL NOT EXCEED THE AMOUNT OF FEES, IF ANY,
+ * THAT YOU HAVE PAID DIRECTLY TO MICROCHIP FOR THIS SOFTWARE.
  *
  * \asf_license_stop
+ *
  */
 
 #ifdef _SAMD51_RSTC_COMPONENT_
@@ -64,83 +55,83 @@ typedef uint8_t hri_rstc_rcause_reg_t;
 
 static inline bool hri_rstc_get_RCAUSE_POR_bit(const void *const hw)
 {
-	return (((Rstc *)(uintptr_t)hw)->RCAUSE.reg & RSTC_RCAUSE_POR) >> RSTC_RCAUSE_POR_Pos;
+	return (((Rstc *)hw)->RCAUSE.reg & RSTC_RCAUSE_POR) >> RSTC_RCAUSE_POR_Pos;
 }
 
 static inline bool hri_rstc_get_RCAUSE_BODCORE_bit(const void *const hw)
 {
-	return (((Rstc *)(uintptr_t)hw)->RCAUSE.reg & RSTC_RCAUSE_BODCORE) >> RSTC_RCAUSE_BODCORE_Pos;
+	return (((Rstc *)hw)->RCAUSE.reg & RSTC_RCAUSE_BODCORE) >> RSTC_RCAUSE_BODCORE_Pos;
 }
 
 static inline bool hri_rstc_get_RCAUSE_BODVDD_bit(const void *const hw)
 {
-	return (((Rstc *)(uintptr_t)hw)->RCAUSE.reg & RSTC_RCAUSE_BODVDD) >> RSTC_RCAUSE_BODVDD_Pos;
+	return (((Rstc *)hw)->RCAUSE.reg & RSTC_RCAUSE_BODVDD) >> RSTC_RCAUSE_BODVDD_Pos;
 }
 
 static inline bool hri_rstc_get_RCAUSE_NVM_bit(const void *const hw)
 {
-	return (((Rstc *)(uintptr_t)hw)->RCAUSE.reg & RSTC_RCAUSE_NVM) >> RSTC_RCAUSE_NVM_Pos;
+	return (((Rstc *)hw)->RCAUSE.reg & RSTC_RCAUSE_NVM) >> RSTC_RCAUSE_NVM_Pos;
 }
 
 static inline bool hri_rstc_get_RCAUSE_EXT_bit(const void *const hw)
 {
-	return (((Rstc *)(uintptr_t)hw)->RCAUSE.reg & RSTC_RCAUSE_EXT) >> RSTC_RCAUSE_EXT_Pos;
+	return (((Rstc *)hw)->RCAUSE.reg & RSTC_RCAUSE_EXT) >> RSTC_RCAUSE_EXT_Pos;
 }
 
 static inline bool hri_rstc_get_RCAUSE_WDT_bit(const void *const hw)
 {
-	return (((Rstc *)(uintptr_t)hw)->RCAUSE.reg & RSTC_RCAUSE_WDT) >> RSTC_RCAUSE_WDT_Pos;
+	return (((Rstc *)hw)->RCAUSE.reg & RSTC_RCAUSE_WDT) >> RSTC_RCAUSE_WDT_Pos;
 }
 
 static inline bool hri_rstc_get_RCAUSE_SYST_bit(const void *const hw)
 {
-	return (((Rstc *)(uintptr_t)hw)->RCAUSE.reg & RSTC_RCAUSE_SYST) >> RSTC_RCAUSE_SYST_Pos;
+	return (((Rstc *)hw)->RCAUSE.reg & RSTC_RCAUSE_SYST) >> RSTC_RCAUSE_SYST_Pos;
 }
 
 static inline bool hri_rstc_get_RCAUSE_BACKUP_bit(const void *const hw)
 {
-	return (((Rstc *)(uintptr_t)hw)->RCAUSE.reg & RSTC_RCAUSE_BACKUP) >> RSTC_RCAUSE_BACKUP_Pos;
+	return (((Rstc *)hw)->RCAUSE.reg & RSTC_RCAUSE_BACKUP) >> RSTC_RCAUSE_BACKUP_Pos;
 }
 
 static inline hri_rstc_rcause_reg_t hri_rstc_get_RCAUSE_reg(const void *const hw, hri_rstc_rcause_reg_t mask)
 {
 	uint8_t tmp;
-	tmp = ((Rstc *)(uintptr_t)hw)->RCAUSE.reg;
+	tmp = ((Rstc *)hw)->RCAUSE.reg;
 	tmp &= mask;
 	return tmp;
 }
 
 static inline hri_rstc_rcause_reg_t hri_rstc_read_RCAUSE_reg(const void *const hw)
 {
-	return ((Rstc *)(uintptr_t)hw)->RCAUSE.reg;
+	return ((Rstc *)hw)->RCAUSE.reg;
 }
 
 static inline bool hri_rstc_get_BKUPEXIT_RTC_bit(const void *const hw)
 {
-	return (((Rstc *)(uintptr_t)hw)->BKUPEXIT.reg & RSTC_BKUPEXIT_RTC) >> RSTC_BKUPEXIT_RTC_Pos;
+	return (((Rstc *)hw)->BKUPEXIT.reg & RSTC_BKUPEXIT_RTC) >> RSTC_BKUPEXIT_RTC_Pos;
 }
 
 static inline bool hri_rstc_get_BKUPEXIT_BBPS_bit(const void *const hw)
 {
-	return (((Rstc *)(uintptr_t)hw)->BKUPEXIT.reg & RSTC_BKUPEXIT_BBPS) >> RSTC_BKUPEXIT_BBPS_Pos;
+	return (((Rstc *)hw)->BKUPEXIT.reg & RSTC_BKUPEXIT_BBPS) >> RSTC_BKUPEXIT_BBPS_Pos;
 }
 
 static inline bool hri_rstc_get_BKUPEXIT_HIB_bit(const void *const hw)
 {
-	return (((Rstc *)(uintptr_t)hw)->BKUPEXIT.reg & RSTC_BKUPEXIT_HIB) >> RSTC_BKUPEXIT_HIB_Pos;
+	return (((Rstc *)hw)->BKUPEXIT.reg & RSTC_BKUPEXIT_HIB) >> RSTC_BKUPEXIT_HIB_Pos;
 }
 
 static inline hri_rstc_bkupexit_reg_t hri_rstc_get_BKUPEXIT_reg(const void *const hw, hri_rstc_bkupexit_reg_t mask)
 {
 	uint8_t tmp;
-	tmp = ((Rstc *)(uintptr_t)hw)->BKUPEXIT.reg;
+	tmp = ((Rstc *)hw)->BKUPEXIT.reg;
 	tmp &= mask;
 	return tmp;
 }
 
 static inline hri_rstc_bkupexit_reg_t hri_rstc_read_BKUPEXIT_reg(const void *const hw)
 {
-	return ((Rstc *)(uintptr_t)hw)->BKUPEXIT.reg;
+	return ((Rstc *)hw)->BKUPEXIT.reg;
 }
 
 #ifdef __cplusplus
